+++ HISTORY.rst
@@ -13,6 +13,8 @@ History
   validation.
 * Added the input ``/event/party``. This is the party submitting the
   transaction.
+* Added the input ``/payment/method``. This is the payment method associated
+  with the transaction.
 
 3.1.0 (2025-05-23)
 ++++++++++++++++++